chore: 셔틀 관련 타입 정의 추가
