+++ src/features/shuttleSchedule/schedule.types.ts
@@ -1,3 +1,5 @@
+import { Passenger } from "@/types/shuttle.types";
+
 export interface ScheduleData {
   [date: string]: ScheduleType;
 }
@@ -15,13 +17,5 @@ export interface BusRoute {
 
 export interface TimeSlot {
   time: string;
-  peopleCount: number;
   passengers?: Passenger[];
-}
-
-export interface Passenger {
-  name: string;
-  contact: string;
-  address: string;
-  boardingLocation: string;
 }
