+++ src/mocks/mockScheduleData.ts.ts
@@ -9,7 +9,6 @@ export const MOCK_SCHEDULE_DATA: ScheduleData = {
         timeSlots: [
           {
             time: "09:00",
-            peopleCount: 3,
             passengers: [
               {
                 name: "홍길동",
@@ -33,7 +32,6 @@ export const MOCK_SCHEDULE_DATA: ScheduleData = {
           },
           {
             time: "11:00",
-            peopleCount: 2,
             passengers: [
               {
                 name: "이상훈",
@@ -51,7 +49,6 @@ export const MOCK_SCHEDULE_DATA: ScheduleData = {
           },
           {
             time: "13:00",
-            peopleCount: 12,
             passengers: [
               {
                 name: "강지원",
@@ -135,7 +132,6 @@ export const MOCK_SCHEDULE_DATA: ScheduleData = {
         timeSlots: [
           {
             time: "09:30",
-            peopleCount: 2,
             passengers: [
               {
                 name: "강지원",
@@ -153,7 +149,6 @@ export const MOCK_SCHEDULE_DATA: ScheduleData = {
           },
           {
             time: "11:30",
-            peopleCount: 3,
             passengers: [
               {
                 name: "허지원",
@@ -185,7 +180,6 @@ export const MOCK_SCHEDULE_DATA: ScheduleData = {
         timeSlots: [
           {
             time: "14:00",
-            peopleCount: 3,
             passengers: [
               {
                 name: "임하진",
@@ -220,7 +214,6 @@ export const MOCK_SCHEDULE_DATA: ScheduleData = {
         timeSlots: [
           {
             time: "08:00",
-            peopleCount: 5,
             passengers: [
               {
                 name: "김철수",
