+++ src/types/shuttle.types.ts
@@ -1,0 +1,8 @@
+export type ShuttleType = "pickup" | "dropoff";
+
+export interface Passenger {
+  name: string;
+  contact: string;
+  address: string;
+  boardingLocation: string;
+}
